+++ packages/web-server/src/index.ts
@@ -19,3 +19,4 @@ export * from './service-description-ref';
 export * from './web-service-compiler';
 export * from './certificate-generator';
 export * from './web-conduit';
+export * from './intercept';