+++ packages/web-server/src/metadata/web-event.ts
@@ -1,6 +1,8 @@
 import * as http from 'http';
 import type { WebServer } from '../web-server';
 import { RouteInstance } from '../route';
+import { InjectionToken } from '@alterior/di';
+import { AnyConstructor } from '@alterior/runtime';
 
 export interface ServerSentEvent {
 	event?: string;
@@ -38,6 +40,14 @@ export class WebEvent<
 	route: RouteInstance;
 
 	requestId: string;
+
+	inject<T>(token: InjectionToken<T> | AnyConstructor<T>): T {
+		return this.server.injector.get(token);
+	}
+
+	static inject<T>(token: InjectionToken<T> | AnyConstructor<T>): T {
+		return this.current.inject(token);
+	}
 
 	/**
 	 * An arbitrary place to store metadata regarding this request.